chore: remove chatActionsImpl
