+++ packages/ai-chat/src/chat/utils/mockStorage.ts
