+++ packages/ai-chat/tests/config/spec/header_spec.ts
@@ -193,7 +193,7 @@ describe("Config Header", () => {
       serviceManager = {
         store,
         namespace: new NamespaceService("test"),
-        messageService: { timeoutMS: 30000 } as any,
+        messageOutboundService: { timeoutMS: 30000 } as any,
         humanAgentService: null,
       } as ServiceManager;
     });
