+++ packages/ai-chat/tests/config/spec/homescreen_spec.ts
@@ -141,7 +141,7 @@ describe("Config Homescreen", () => {
         serviceManager = {
           store,
           namespace: new NamespaceService("test"),
-          messageService: { timeoutMS: 30000 } as any,
+          messageOutboundService: { timeoutMS: 30000 } as any,
           humanAgentService: null,
         } as ServiceManager;
       });
