+++ packages/ai-chat/tests/config/spec/launcher_spec.ts
@@ -209,7 +209,7 @@ describe("Config Launcher", () => {
       serviceManager = {
         store,
         namespace: new NamespaceService("test"),
-        messageService: { timeoutMS: 30000 } as any,
+        messageOutboundService: { timeoutMS: 30000 } as any,
         humanAgentService: null,
       } as ServiceManager;
     });
