+++ packages/ai-chat/tests/config/spec/layout_spec.ts
@@ -174,7 +174,7 @@ describe("Config Layout", () => {
         serviceManager = {
           store,
           namespace: new NamespaceService("test"),
-          messageService: { timeoutMS: 30000 } as any,
+          messageOutboundService: { timeoutMS: 30000 } as any,
           humanAgentService: null,
         } as ServiceManager;
       });
