+++ packages/ai-chat/tests/config/spec/misc_spec.ts
@@ -236,7 +236,7 @@ describe("Config Miscellaneous", () => {
         serviceManager = {
           store,
           namespace: new NamespaceService("test"),
-          messageService: { timeoutMS: 30000 } as any,
+          messageOutboundService: { timeoutMS: 30000 } as any,
           humanAgentService: null,
         } as ServiceManager;
       });
