+++ packages/ai-chat/tests/config/spec/namespace_spec.ts
@@ -163,7 +163,7 @@ describe("Config Namespace", () => {
         serviceManager = {
           store,
           namespace: new NamespaceService("test"),
-          messageService: { timeoutMS: 30000 } as any,
+          messageOutboundService: { timeoutMS: 30000 } as any,
           humanAgentService: null,
         } as ServiceManager;
       });
